Adding to flow generate face thumbnails

+++ src/flows/cleanup.py
@@ -1,6 +1,7 @@
 from prefect import flow, task
 from dotenv import load_dotenv
 import os
+import shutil
 
 load_dotenv()  # Inject environment variables from .env during development
         
@@ -9,8 +10,14 @@ def cleanup():
     """
     Remove all generated files
     """
-    os.remove(os.environ["DATABASE_PATH"])
-    os.remove(os.environ["EMBEDDINGS_INDEX_PATH"])
+    if os.path.exists(os.environ["DATABASE_PATH"]):
+        os.remove(os.environ["DATABASE_PATH"])
+        
+    if os.path.exists(os.environ["EMBEDDINGS_INDEX_PATH"]):
+        os.remove(os.environ["EMBEDDINGS_INDEX_PATH"])
+        
+    if os.path.exists(os.environ["THUMBNAILS_PATH"]):
+        shutil.rmtree(os.environ["THUMBNAILS_PATH"])
     
 if __name__ == "__main__":
     cleanup()
